Extract ChartState accessors and methods into state.rs
Move accessors, setters, computed properties, and instance methods from
chart/mod.rs (1061 lines) into chart/state.rs (367 lines), bringing
mod.rs down to 712 lines. Both files are now well under the 1000-line
limit, leaving room for the cursor/crosshair feature.
